Add unit tests for the login controller using Jest
