+++ models/user.js
@@ -31,7 +31,6 @@ const userSchema = new Schema({
     },
     avatarUrl: {
         type: String,
-        required: true,
     },
 }, {versionKey: false, timestamps: true});
